+++ src/main/java/bowling/domain/dto/BowlingGameDto.java
@@ -1,0 +1,40 @@
+package bowling.domain.dto;
+
+import java.util.List;
+
+public class BowlingGameDto {
+
+    private String playerName;
+    private List<ScoreSignaturesDto> scoreSignatures;
+    private List<Integer> scores;
+
+    public BowlingGameDto(String playerName, List<ScoreSignaturesDto> scoreSignatures, List<Integer> scores) {
+        this.playerName = playerName;
+        this.scoreSignatures = scoreSignatures;
+        this.scores = scores;
+    }
+
+    public String getPlayerName() {
+        return playerName;
+    }
+
+    public void setPlayerName(String playerName) {
+        this.playerName = playerName;
+    }
+
+    public List<ScoreSignaturesDto> getScoreSignatures() {
+        return scoreSignatures;
+    }
+
+    public void setScoreSignatures(List<ScoreSignaturesDto> scoreSignatures) {
+        this.scoreSignatures = scoreSignatures;
+    }
+
+    public List<Integer> getScores() {
+        return scores;
+    }
+
+    public void setScores(List<Integer> scores) {
+        this.scores = scores;
+    }
+}
